Extracted template from filter component
