+++ src/components/filter/index.js
@@ -1,0 +1,8 @@
+import Component from "../component";
+import {template} from "./template";
+
+export default class Filter extends Component {
+  constructor(filters) {
+    super(template(filters));
+  }
+}
